+++ src/main/java/com/ksu/soccerserver/comment/CommentController.java
@@ -2,10 +2,13 @@ package com.ksu.soccerserver.comment;
 
 import com.ksu.soccerserver.account.Account;
 import com.ksu.soccerserver.account.AccountRepository;
+import com.ksu.soccerserver.account.CurrentAccount;
 import com.ksu.soccerserver.board.Board;
 import com.ksu.soccerserver.board.BoardRepository;
 import com.ksu.soccerserver.comment.dto.CommentRequest;
+import com.ksu.soccerserver.comment.dto.CommentResponse;
 import lombok.RequiredArgsConstructor;
+import org.modelmapper.ModelMapper;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -15,60 +18,74 @@ import org.springframework.web.server.ResponseStatusException;
 import java.time.LocalDateTime;
 import java.util.List;
 
-@RequestMapping("/api/boards")
+@RequestMapping("/api/comments")
 @RequiredArgsConstructor
 @RestController
 public class CommentController {
     private final CommentRepository commentRepository;
     private final AccountRepository accountRepository;
     private final BoardRepository boardRepository;
+    private final ModelMapper modelMaapper;
 
-    @PostMapping("/{boardId}/comment/{accountId}")
-    ResponseEntity<?> postComment(@RequestBody CommentRequest commentRequest, @PathVariable Long boardId, @PathVariable Long accountId) {
+    @PostMapping
+    ResponseEntity<?> postComment(@RequestBody CommentRequest commentRequest, @CurrentAccount Account currentAccount) {
 
-        Account account = accountRepository.findById(accountId).get();
-        Board board = boardRepository.findById(boardId).get();
-        Comment saveComment = commentRepository.save(Comment.builder()
-                .content(commentRequest.getContent())
-                .createdAt(LocalDateTime.now())
-                .build());
+        Account account = accountRepository.findById(currentAccount.getId())
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND,"NO_FOUND_ACCOUNT"));
 
-        saveComment.commentAccount(account);
-        saveComment.commentBoard(board);
+        Board board = boardRepository.findById(commentRequest.getBoardId())
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND,"NO_FOUND_BOARD"));
 
-        commentRepository.save(saveComment);
-        return new ResponseEntity<>("Create new Comments" + board.getId() + "번째 게시판", HttpStatus.CREATED);
+        Comment postComment = commentRequest.toEntity(board, account);
+        Comment saveComment = commentRepository.save(postComment);
+
+        CommentResponse response = modelMaapper.map(saveComment, CommentResponse.class);
+
+        return new ResponseEntity<>(response,HttpStatus.CREATED);
     }
 
+    /*
     @GetMapping("/{boardId}/comment")
     ResponseEntity<?> getComment(@PathVariable Long boardId){
         List<Comment> comments = commentRepository.findByBoard(boardRepository.findById(boardId).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "존재하지 않는 게시판입니다.")));
         if(comments.isEmpty()){
-            return new ResponseEntity<>("댓글이 없습니다.", HttpStatus.NOT_FOUND);
+            return new ResponseEntity<>("댓글이 없습니다.", HttpStatus.NOT_FOUND);.x
         }
         return new ResponseEntity<>(comments, HttpStatus.OK);
     }
+    */
 
-    @PutMapping("/{boardId}/comment/{commentId}")
-    ResponseEntity<?> putComment(@RequestBody CommentRequest commentRequest, @PathVariable Long boardId, @PathVariable Long commentId){
-        Board board = boardRepository.findById(boardId).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "존재하지 않는 게시판입니다."));
+    @PutMapping("/{commentId}")
+    ResponseEntity<?> putComment(@RequestBody CommentRequest commentRequest, @PathVariable Long commentId, @CurrentAccount Account currentAccount){
         Comment findComment = commentRepository.findById(commentId).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "존재하지 않는 댓글입니다."));
+        if(!currentAccount.getId().equals(findComment.getAccount().getId())){
+            return new ResponseEntity<>("수정권한이 없습니다.", HttpStatus.BAD_REQUEST);
+        }else {
 
-        findComment.setContent(commentRequest.getContent());
-        findComment.setTime(LocalDateTime.now());
+            findComment.setContent(commentRequest.getContent());
+            findComment.setTime(LocalDateTime.now());
 
-        commentRepository.save(findComment);
-        return new ResponseEntity<>(findComment, HttpStatus.OK);
+            Comment updatedComment = commentRepository.save(findComment);
+
+            CommentResponse response = modelMaapper.map(updatedComment, CommentResponse.class);
+            return new ResponseEntity<>(response, HttpStatus.OK);
+        }
 
     }
 
-    @DeleteMapping("/{boardId}/comment/{commentId}")
-    ResponseEntity<?> deleteComment(@PathVariable Long boardId, @PathVariable Long commentId){
-        Board board = boardRepository.findById(boardId).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "존재하지 않는 게시판입니다."));
+    @DeleteMapping("/{commentId}")
+    ResponseEntity<?> deleteComment(@PathVariable Long commentId,@CurrentAccount Account currentAccount){
         Comment comment = commentRepository.findById(commentId).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "존재하지 않는 댓글입니다."));
+        if(!currentAccount.getId().equals(comment.getAccount().getId())){
+            return new ResponseEntity<>("삭제권한이 없습니다.",HttpStatus.BAD_REQUEST);
+        }else {
 
-        commentRepository.delete(comment);
-        return new ResponseEntity<>(board.getTitle()+"의 댓글 Delete : " + comment.getContent(), HttpStatus.OK );
+            commentRepository.delete(comment);
+
+            CommentResponse response = modelMaapper.map(comment, CommentResponse.class);
+
+            return new ResponseEntity<>(response, HttpStatus.OK);
+        }
     }
